Fix SwiftyBeaver warnings in BunPuroKit & more
