+++ BunPuro/Status/StatusTableViewController.swift
@@ -408,55 +408,3 @@ extension StatusTableViewController: SFSafariViewControllerDelegate {
         dismiss(animated: true, completion: nil)
     }
 }
-
-extension Collection where Iterator.Element == Review {
-    public var nextReviewDate: Date? {
-        let allDates = filter { $0.complete }.compactMap { $0.nextReviewDate }
-
-        let tmp = allDates.reduce(Date.distantFuture) { $0 < $1 ? $0 : $1 }
-        return tmp == Date.distantPast ? nil: tmp
-    }
-
-    public func reviews(at date: Date) -> [Review] {
-        let result = filter { $0.complete && $0.nextReviewDate!.minutes(from: date) <= 0 }
-        return result
-    }
-
-    public var reviewsWithinNextHour: Int {
-        let date = Date()
-        let result = filter { $0.complete && $0.nextReviewDate!.minutes(from: date) < 60 }
-        return result.count
-    }
-
-    public var reviewsWithNext24Hours: Int {
-        let date = Date()
-        let result = filter { $0.complete && $0.nextReviewDate!.hours(from: date) <= 23 }
-        return result.count
-    }
-}
-
-extension Date {
-    func minutes(from date: Date) -> Int {
-        return Calendar.autoupdatingCurrent.dateComponents([.minute], from: date, to: self).minute!
-    }
-
-    func hours(from date: Date) -> Int {
-        return Calendar.autoupdatingCurrent.dateComponents([.hour], from: date, to: self).hour!
-    }
-
-    func isTomorrow() -> Bool {
-        return Calendar.autoupdatingCurrent.isDateInTomorrow(self)
-    }
-
-    var tomorrow: Date {
-        return Calendar.autoupdatingCurrent.date(byAdding: .day, value: 1, to: noon)!
-    }
-
-    var noon: Date {
-        return Calendar.autoupdatingCurrent.date(bySettingHour: 12, minute: 0, second: 0, of: self)!
-    }
-
-    var nextMidnight: Date {
-        return Calendar.autoupdatingCurrent.date(bySettingHour: 0, minute: 0, second: 0, of: self)!
-    }
-}
