+++ BunPuroKit/Logger.swift
@@ -1,0 +1,8 @@
+//
+//  Created by Cihat Gündüz on 22.04.19.
+//  Copyright © 2019 Andreas Braun. All rights reserved.
+//
+
+import SwiftyBeaver
+
+let log = SwiftyBeaver.self
