+++ BunPuroKit/Procedures/StatusProcedure.swift
@@ -45,10 +45,10 @@ public final class StatusProcedure: GroupProcedure, OutputProcedure {
             if let decodingError = error as? Swift.DecodingError {
                 switch decodingError {
                 case .typeMismatch, .valueNotFound:
-                    print("Server changed something!")
+                    BunPuroKit.log.info("Server changed something!")
 
                 case .keyNotFound:
-                    print("Login token changed")
+                    BunPuroKit.log.info("Login token changed")
                     Server.reset()
 
                 case .dataCorrupted:
